Add files via upload
This code takes in a trace file line by line and does the following:
stores n in a global integer variable
stores the address data in a global 8 element array of integers
breaks the address down into its tag, index and byte offset and stores these into a single declared struct
Currently the code prints each of the above to the terminal and replaces each actual value in the global variables/struct with each line read from the trace file. We will need to make sure our cache simulation performs the appropriate action with each n and address before moving on to the next line in the trace file.

It is assumed that the trace file will be named "TraceFile.txt" We will need to confirm with Yuchen/Sonali that this is correct before submitting. If this is not correct, we will need to change line 47 to reflect the appropriate file name. 

The code accounts for several possible errors found in the trace file:
1.) If the file fails to open an error message is  printed to the terminal and the program exits.
2.) If a line contains no hex values (including completely blank lines, lines with just horizontal whitespace, lines with non hex values, or any and all combinations of these three) that line will be ignored by the program and no data will be generated.
3.)  If a line contains hex values it will be properly read and stored and ignore all horizontal whitespace and non hex values.
4.) if there are more than 9 hex elements on a single line (1 for n, 8 for address) the following hex values will be ignored.
5.) If a non hex value were to inexplicably slip past the prior checks, if a non hex value is found when converting from a character to an integer the program will print an error message to the terminal and the program will terminate.
6.) If an invalid number is given to n (that is 5,6,7, >9) an error is printed and the program terminates.
7.) If fewer than 8 hex values are given in the address, the program will give the appropriate number of leading zeros to fill out the address to 8 hex values.
8.) An error is printed to the terminal if the file fails to properly close.
 
The program cannot handle the following:
1.) If a given line in the trace file has relevant n or address data past 999 characters, it will be lost. This should not be a problem.
2.) If the n or address data is split across multiple lines, the program will not be able to identify this.
3.) Similar to 2. above, if a line contains multiple n address bundles, only the first one will be utilized, the rest will be ignored.
Both 2. and 3. above should not be a problem and are impossible to solve for. Good data must be entered into the Trace File.

I talked to Yuchen and a lot of these features were not necessary, but honestly this was driving me up the wall and I am relieved to say I finally got it working. One nice thing is that if a typo is made into the trace file it is very likely our program will be able to handle and produce the correct results around it.